services/incident-classifier/requirements.txt
kafka-python
PyYAML